First version
I was inspired by a video from Nick Chapsas to explore the possibility of creating a lightweight C# object extension that could offer performance and memory efficiency comparable to hand-coding the mapping or using a code generator. My objective was to create a solution that relied solely on simple .NET, with no additional packages or configuration needed. The result is a straightforward, dynamic, and lightweight solution that I have built. I'd like to share this implementation, hoping Nick may consider benchmarking it against the packages he has been testing. With its simplicity and minimalistic approach, I believe this solution is worth considering. What are your thoughts, Nick?

The Best .NET Mapper to Use in 2023 by Nick Chapsas
https://youtu.be/U8gSdQN2jWI